ISPN-7506 Tests fail after upgrading Lucene version
