+++ integrationtests/wildfly-modules/src/test/java/org/infinispan/test/integration/as/query/ElasticsearchIndexManagerIT.java
@@ -1,11 +1,7 @@
 package org.infinispan.test.integration.as.query;
 
-import org.elasticsearch.plugin.deletebyquery.DeleteByQueryPlugin;
 import org.infinispan.Version;
-import org.infinispan.query.test.elasticsearch.ElasticSearchCluster;
-import org.infinispan.query.test.elasticsearch.ElasticSearchCluster.ElasticSearchClusterBuilder;
 import org.jboss.arquillian.container.test.api.Deployment;
-import org.jboss.arquillian.container.test.api.RunAsClient;
 import org.jboss.arquillian.container.test.api.TargetsContainer;
 import org.jboss.arquillian.junit.Arquillian;
 import org.jboss.shrinkwrap.api.Archive;
@@ -16,8 +12,6 @@ import org.jboss.shrinkwrap.api.asset.StringAsset;
 import org.jboss.shrinkwrap.api.spec.WebArchive;
 import org.jboss.shrinkwrap.descriptor.api.Descriptors;
 import org.jboss.shrinkwrap.descriptor.api.spec.se.manifest.ManifestDescriptor;
-import org.junit.AfterClass;
-import org.junit.BeforeClass;
 import org.junit.runner.RunWith;
 
 /**
@@ -40,7 +34,7 @@ public class ElasticsearchIndexManagerIT extends BaseQueryTest {
 
    private static Archive<?> deployment() {
       return ShrinkWrap.create(WebArchive.class)
-            .addClasses(ElasticSearchCluster.class, BaseQueryTest.class, ElasticsearchIndexManagerIT.class,
+            .addClasses(BaseQueryTest.class, ElasticsearchIndexManagerIT.class,
                   ElasticQueryConfiguration.class, Book.class, GridService.class)
             .add(manifest(), "META-INF/MANIFEST.MF")
             .addAsResource("elasticsearch-indexing.xml")
@@ -56,25 +50,5 @@ public class ElasticsearchIndexManagerIT extends BaseQueryTest {
       return new StringAsset(manifest);
    }
 
-
-   private static ElasticSearchCluster elasticSearchCluster;
-
-   @BeforeClass
-   @RunAsClient
-   public static void startElastic() throws Exception {
-      elasticSearchCluster = new ElasticSearchClusterBuilder()
-            .withNumberNodes(2)
-            .addPlugin(DeleteByQueryPlugin.class)
-            .build();
-      elasticSearchCluster.start();
-   }
-
-   @AfterClass
-   @RunAsClient
-   public static void stopElastic() throws Exception {
-      if (elasticSearchCluster != null) {
-         elasticSearchCluster.stop();
-      }
-   }
 
 }
