+++ query/pom.xml
@@ -151,18 +151,8 @@
 	 <scope>test</scope>
       </dependency>
       <dependency>
-         <groupId>org.elasticsearch</groupId>
-         <artifactId>elasticsearch</artifactId>
-         <scope>test</scope>
-      </dependency>
-      <dependency>
          <groupId>log4j</groupId>
          <artifactId>log4j</artifactId>
-         <scope>test</scope>
-      </dependency>
-      <dependency>
-         <groupId>org.elasticsearch.plugin</groupId>
-         <artifactId>delete-by-query</artifactId>
          <scope>test</scope>
       </dependency>
    </dependencies>
